Update association_basics.md
Moves the definition of an association to the top of the page. I am just starting to learn Rails, and having this definition at the top instead of in the second section would be really useful.

Updates the Types of Associations section for clarity. Moves the list of associations before the explanation.

Links to wikipedia articles on Primary and Foreign keys.