base (refined) Pathname#ascendant_of? also on Pathname#ascend
A small rewrite in a last attempt at writing obvious and portable code
without manual string manipulation.

Note that Pathname#== uses string comparison on Windows, so if client
code passes "C:\foo" and "c:/foo/bar" the predicate won't see the
former is an ascendant of the latter. Risky business.