guides, scaffold no longer used in getting started guide. Closes #22337.
[ci skip]

Since the "Getting Started" guide no longer uses the scaffold generator
we should rewrite references to that in the testing guide.

The functional testing section was quite heavily based on such a
scaffold test. I changed it to use `generate scaffold_controller`
instead so that we can build up on the model foundation we already have.