Add class level `run` to Rails command.
The class level version is responsible for changing a task name to
command name, then finding a command and run it if there is one.

The instance level `run` then makes sure arguments have been parsed
into `@options` and runs the command by sending it.

`Rails::Commands::Command.run` returns true to make it work within
`Rails::CommandsTask`, but won't in the future when it handles
all option parsing.