Remove "run with rake" suggestion.
When attempting to run a `rake` task with `bin/rails`, Rails would see
check that a Rake task existed and ask users if they meant to run it with
rake.

We don't need that anymore, as the goal is to have `rails` be a valid way
to run rake tasks.