Ensure we install the database backend gems on Windows
platform `:ruby` is only MRI on Linux or Mac. Windows MRI shows up as
`:mswin` or `:mswin64` depending on if it was installed as 32 or 64 bit. I
am unsure if this will cause conflicts with JRuby on Windows, but I don't
have the means to test this at the moment.